Fix: run 'apt-get update' before installing grub
We may have unpacked the rootfs tarball, and the Packages files may be
out of date. The 'apt-get update' fixes this, and is not too slow.

This is a crummy fix. Better would be for the debootstrap step to
always run the update, but the way steps are skipped doesn't allow
that at the moment.