Change: use subprocess directly, not via cliapp.runcmd
We didn't use the cliapp.runcmd funcationality at all, except for the
stdout/stderr callbacks, and those only to log the output as it
happens. The logs don't seem all that useful, especially in real time.
We add back the logging of command output if it turns out there's a
need for that.