refactor(plugins): replace cliapp's plugin infra with a custom one
Previously, we were using the plugin system in the cliapp library.
However, as we want to drop depenendency on cliapp, we can't do that
in the future. This commit adds a new plugin system. Because all the
plugins (vmdb/plugins/*.py) need to be changed for the new system, the
diff is pretty long, but most of it is just obvious mechanical changes.