Fix logic of the apt plugin's 'recommends' option
The documentation says to set 'recommends' to true or false, with false
telling apt not to install recommends. However, the logic in the code does
the opposite of that. This also means that by default, all apt-get install
calls from vmdb2 run with --no-install-recommends.

As a user, it doesn't make much sense to set 'recommends' to true to tell
apt not to install recommends. So I think fixing the logic in the code
makes more sense than updating the documentation.