grub_plugin: add support for the root filesystem on a logical volume.
Detect if the root filesystem is inside a know logical volume group by
examining the output of the `lvs` command. If the rootfs is in a known
LV group, determine the name of the mapped device for the group. If the
`lvs` command shows the device as `/dev/dm-X` instead of
`/dev/mapper/loopXpY`, use the `dmsetup ls` command to determine the
name of the loop device.