Fix part-naming and add timeout while waiting
When installing on a blockdevice which device name ends in a number,
a 'p' is added between the device name and the partition number.
This will be the case, for example, when installing to /dev/nbd0 or
/dev/nvme0.

Further, added a timeout to the wait_for_file_to_exist function,
in order not to hang when a partition was not found.

Both changes should resolve #46